# tp-spring/openapi.yaml
openapi: 3.0.3
info:
  title: TP Web INSA Rennes
  description: |
    Introduction to REST back-end development in Java with Spring and OpenAPI
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v2
tags:
  - name: todo
    description: REST operations on Todos
  - name: todolist
    description: REST operations on Todo lists
paths:
  /public/todo/hello:
    get:
      tags:
        - todo
      responses:
        '200':
          description: Success
  /public/todo/todo:
    get:
      tags:
        - todo
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    post:
      tags:
        - todo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
    put:
      tags:
        - todo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
        required: true
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid ID supplied
  /public/todo/todo/{id}:
    patch:
      tags:
        - todo
      parameters:
        - name: id
          in: path
          description: Identifier of the todo to patch
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Todo'
        required: true
      responses:
        '200':
          description: Success
        '400':
          description: Invalid ID supplied
    delete:
      tags:
        - todo
      parameters:
        - name: id
          in: path
          description: Identifier of the todo to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied

  /public/todo/title:
    get:
      tags:
        - todo
      parameters:
        - name: txt
          in: query
          description: Text fragment to search for in titles
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'

  /public/todolist:
    post:
      tags:
        - todolist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NamedDTO'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoList'
  /public/todolist/addtodo:
    patch:
      description: Add a todo into a todo list (using a DTO to provide ids)
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IdsDTO'
        required: true
      tags:
        - todolist
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoList'
  /public/todolist/{todoListId}/addtodo/{todoId}:
    patch:
      description: Add a todo into a todo list (using path variables to provide ids)
      parameters:
        - name: todoListId
          in: path
          description: Identifier of the todo list
          required: true
          schema:
            type: integer
            format: int64
        - name: todoId
          in: path
          description: Identifier of the todo
          required: true
          schema:
            type: integer
            format: int64
      tags:
        - todolist
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoList'

components:
  schemas:
    Todo:
      type: object
      properties:
        type:
          type: string
          enum:
            - todo
            - spec
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        categories:
          type: array
          items:
            type: string
            enum:
              - HIGH_PRIORITY
              - LOW_PRIORITY
              - WORK
              - ENTERTAINMENT
    TodoList:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        owner:
          type: string
        todos:
          type: array
          items:
            $ref: '#/components/schemas/Todo'
    NamedDTO:
      type: object
      properties:
        name:
          type: string
    IdsDTO:
      type: object
      properties:
        todoListId:
          type: integer
          format: int64
        todoId:
          type: integer
          format: int64